ajout de check

+++ setup/setup.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 
+set -e  # Exit immediately if a command exits with a non-zero status
+
 # Update and install prerequisites
 sudo apt-get update
 sudo apt-get install -y cmake
+
+# Check if conda is available
+if ! command -v conda &> /dev/null; then
+    echo "Conda could not be found, please install Miniconda or Anaconda."
+    exit 1
+fi
 
 # Create conda environment
 conda create -n flower_detection python=3.11 -y
@@ -16,7 +24,12 @@ sudo apt install -y python3-libcamera python3-kms++
 sudo apt install -y python3-pyqt5 python3-prctl libatlas-base-dev ffmpeg python3-pip
 
 # Install Python packages from requirements.txt
-pip install -r setup/requirements.txt
+if [ -f setup/requirements.txt ]; then
+    pip install -r setup/requirements.txt
+else
+    echo "requirements.txt not found."
+    exit 1
+fi
 
 # Copy necessary libraries to conda environment
 sudo cp -r /usr/lib/python3/dist-packages/libcamera ~/miniconda3/envs/flower_detection/lib/python3.11/site-packages/
@@ -25,7 +38,7 @@ sudo cp -r /usr/lib/python3/dist-packages/pykms ~/miniconda3/envs/flower_detecti
 # Replace libstdc++ with the system version
 cd ~/miniconda3/envs/flower_detection/lib
 mv -vf libstdc++.so.6 libstdc++.so.6.old
-ln -s /usr/lib/x86_64-linux-gnu/libstdc++.so.6 ./libstdc++.so.6
+ln -s /usr/lib/aarch64-linux-gnu/libstdc++.so.6 ./libstdc++.so.6
 
 # Install QT5
 conda install pyqt -y
@@ -34,4 +47,4 @@ conda install pyqt -y
 pip uninstall -y opencv-python
 pip install opencv-python-headless==4.6.0.66
 
-echo Setup complete. Activate the environment using 'conda activate flower_detection' and run your scripts.
+echo "Setup complete. Activate the environment using 'conda activate flower_detection' and run your scripts."
